Time: 3 ms (25.00%) | Memory: 9.2 MB (30.09%) - LeetSync

+++ 1065-binary-string-with-substrings-representing-1-to-n/binary-string-with-substrings-representing-1-to-n.cpp
@@ -1,7 +1,7 @@
 #define ll long long int
 class Solution {
 public:
-    unordered_map<ll,bool>mp;
+    unordered_map<int,bool>mp;
     bool queryString(string s, int n) {
         int N=s.length(),num=0;
         mp.clear();
